+++ widowx_envs/widowx_envs/multicam_server/launch/camera_cpluplus.launch
