+++ widowx_envs/widowx_envs/multicam_server/src/camera_recorder.py
@@ -231,7 +231,7 @@ class CameraRecorder:
 
 
 if __name__ == '__main__':
-    from widowx_envs.utils.multicam_server_rospkg.src.topic_utils import IMTopic
+    from widowx_envs.multicam_server.src.topic_utils import IMTopic
     rospy.init_node("camera_rec_test")
     imtopic = IMTopic('/cam1/image_raw')
     rec = CameraRecorder(imtopic)
