+++ widowx_envs/widowx_envs/multicam_server/src/usb_connector_chart_example.yml
