--- widowx_envs/widowx_envs/utils/multicam_server_rospkg/launch/launch_default_imgsize.txt
@@ -1,1 +1,0 @@
-roslaunch multicam_server camera_cpluplus.launch video_stream_provider:=0 camera_name:=cam0 visualize:=True fps:=20 buffer_queue_size:=0  width:=0 height:=0
