--- widowx_envs/widowx_envs/utils/video_stream_opencv/.gitignore
@@ -1,4 +1,0 @@
-*~
-*.pyc
-*.mp4
-!test/small.mp4
