--- widowx_envs/widowx_envs/utils/video_stream_opencv/.travis.yml
@@ -1,11 +1,0 @@
-language: generic
-services:
-  - docker
-env:
-  matrix:
-    - ROS_DISTRO="kinetic"
-    - ROS_DISTRO="melodic"
-install:
-  - git clone --quiet --depth 1 https://github.com/ros-industrial/industrial_ci.git .industrial_ci
-script:
-  - ".industrial_ci/travis.sh"
