--- widowx_envs/widowx_envs/utils/video_stream_opencv/launch/rtsp_stream.launch
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-   <!-- launch video stream -->
-   <include file="$(find video_stream_opencv)/launch/camera.launch" >
-   		<!-- node name and ros graph name -->
-	  	<arg name="camera_name" value="rtsp2" />
-	  	<!-- url of the video stream -->
-	  	<arg name="video_stream_provider" value="rtsp://wowzaec2demo.streamlock.net/vod/mp4:BigBuckBunny_115k.mov" />
-	  	<!-- set camera fps to (does nothing on a stream)-->
-	  	<!-- <arg name="set_camera_fps" value="30"/> -->
-      	<!-- set buffer queue size of frame capturing to -->
-      	<arg name="buffer_queue_size" value="1000" />
-	  	<!-- throttling the querying of frames to -->
-	  	<arg name="fps" value="30" />
-	  	<!-- setting frame_id -->
-	  	<arg name="frame_id" value="rtsp_frame" />
-	  	<!-- camera info loading, take care as it needs the "file:///" at the start , e.g.:
-	  	"file:///$(find your_camera_package)/config/your_camera.yaml" -->
-	  	<arg name="camera_info_url" value="" />
-	  	<!-- flip the image horizontally (mirror it) -->
-	  	<arg name="flip_horizontal" value="false" />
-	  	<!-- flip the image vertically -->
-	  	<arg name="flip_vertical" value="false" />
-	  	<!-- visualize on an image_view window the stream generated -->
-	  	<arg name="visualize" value="true" />
-   </include>
-</launch>
