--- widowx_envs/widowx_envs/utils/video_stream_opencv/launch/test_camera_info.launch
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-   <!-- launch video stream -->
-   <include file="$(find video_stream_opencv)/launch/camera.launch" >
-   		<!-- node name and ros graph name -->
-	  	<arg name="camera_name" value="rtsp" />
-	  	<!-- url of the video stream -->
-	  	<arg name="video_stream_provider" value="rtsp://wowzaec2demo.streamlock.net/vod/mp4:BigBuckBunny_115k.mov" />
-	  	<!-- set camera fps to -->
-	  	<arg name="set_camera_fps" value="30"/>
-      	<!-- set buffer queue size of frame capturing to -->
-      	<arg name="buffer_queue_size" value="10" />
-	  	<!-- throttling the querying of frames to -->
-	  	<arg name="fps" value="30" />
-	  	<!-- setting frame_id -->
-	  	<arg name="frame_id" value="rtsp" />
-	  	<!-- camera info loading, take care as it needs the "file:///" at the start , e.g.:-->
-	  	<arg name="camera_info_url" value="file:///$(find video_stream_opencv)/config/test_calibration.yaml"/>
-	  	<!-- visualize on an image_view window the stream generated -->
-	  	<arg name="visualize" value="true" />
-   </include>
-</launch>
