--- widowx_envs/widowx_envs/utils/video_stream_opencv/launch/video_file.launch
@@ -1,34 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-   <!-- launch video stream -->
-   <include file="$(find video_stream_opencv)/launch/camera.launch" >
-        <!-- node name and ros graph name -->
-        <arg name="camera_name" value="videofile" />
-        <!-- full path to the video file -->
-        <!-- wget http://techslides.com/demos/sample-videos/small.mp4 -O /tmp/small.mp4 -->
-        <arg name="video_stream_provider" value="/tmp/small.mp4" />
-        <!-- set camera fps to (video files not affected) -->
-        <!-- <arg name="set_camera_fps" value="30"/> -->
-        <!-- set buffer queue size of frame capturing to -->
-        <arg name="buffer_queue_size" value="1000" />
-        <!-- throttling the querying of frames to -->
-        <arg name="fps" value="30" />
-        <!-- setting frame_id -->
-        <arg name="frame_id" value="videofile_frame" />
-        <!-- camera info loading, take care as it needs the "file:///" at the start , e.g.:
-        "file:///$(find your_camera_package)/config/your_camera.yaml" -->
-        <arg name="camera_info_url" value="" />
-        <!-- flip the image horizontally (mirror it) -->
-        <arg name="flip_horizontal" value="false" />
-        <!-- flip the image vertically -->
-        <arg name="flip_vertical" value="false" />
-        <!-- enable looping playback -->
-        <arg name="loop_videofile" value="true" />
-        <!-- start frame of video -->
-        <arg name="start_frame" default="0"/>
-        <!-- stop frame of video, -1 means the end of video -->
-        <arg name="stop_frame" default="-1"/>
-        <!-- visualize on an image_view window the stream generated -->
-        <arg name="visualize" value="true" />
-   </include>
-</launch>
