--- widowx_envs/widowx_envs/utils/video_stream_opencv/test/test_mjpg_stream.test
@@ -1,47 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-  <arg name="camera_name" default="mjpg" />
-  <arg name="video_stream_provider" default="http://localhost:8080/video.mjpg" />
-  <arg name="set_camera_fps" default="30"/>
-  <arg name="buffer_queue_size" default="100" />
-  <arg name="fps" default="30" />
-  <arg name="frame_id" default="mjpg_frame" />
-  <arg name="camera_info_url" default="" />
-  <arg name="flip_horizontal" default="false" />
-  <arg name="flip_vertical" default="false" />
-  <arg name="loop_videofile" default="false" />
-  <arg name="width" default="0"/>
-  <arg name="height" default="0"/>
-
-  <group ns="$(arg camera_name)">
-    <node pkg="video_stream_opencv" type="video_stream" name="$(arg camera_name)_stream"
-          output="screen" respawn="true">
-      <remap from="camera" to="image_raw" />
-      <param name="camera_name" type="string" value="$(arg camera_name)" />
-      <param name="video_stream_provider" type="string" value="$(arg video_stream_provider)" />
-      <param name="set_camera_fps" type="double" value="$(arg set_camera_fps)" />
-      <param name="buffer_queue_size" type="int" value="$(arg buffer_queue_size)" />
-      <param name="fps" type="double" value="$(arg fps)" />
-      <param name="frame_id" type="string" value="$(arg frame_id)" />
-      <param name="camera_info_url" type="string" value="$(arg camera_info_url)" />
-      <param name="flip_horizontal" type="bool" value="$(arg flip_horizontal)" />
-      <param name="flip_vertical" type="bool" value="$(arg flip_vertical)" />
-      <param name="loop_videofile" type="bool" value="$(arg loop_videofile)" />
-      <param name="width" type="int" value="$(arg width)" />
-      <param name="height" type="int" value="$(arg height)" />
-    </node>
-  </group>
-
-  <node name="mjpg_server" pkg="video_stream_opencv" type="mjpg_server.py"
-        args="--port 8080 $(find video_stream_opencv)/test/small.mp4"
-        output="screen"/>
-
-  <test test-name="pubtest" pkg="rostest" type="publishtest">
-    <rosparam>
-      topics:
-      - name: /mjpg/image_raw
-      - name: /mjpg/camera_info
-      timeout: 10
-    </rosparam>
-  </test>
-</launch>
