+++ widowx_envs/widowx_envs/widowx/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>robonetv2</name>
+  <name>widowx</name>
   <version>0.0.0</version>
   <description>Package for controlling WidowX robots</description>
