+++ widowx_envs/widowx_envs/widowx/src/custom_gripper_controller.py
@@ -5,8 +5,8 @@ from sensor_msgs.msg import JointState
 import time
 
 from std_msgs.msg import Float64
-from robonetv2.srv import OpenGripper, OpenGripperResponse
-from robonetv2.srv import GetGripperDesiredState, GetGripperDesiredStateResponse
+from widowx.srv import OpenGripper, OpenGripperResponse
+from widowx.srv import GetGripperDesiredState, GetGripperDesiredStateResponse
 
 try:
     # older version of interbotix sdk
